Add anisotropic Debye-Waller factor support to Bloch wave module
Anisotropic thermal displacements (sx, sy, sz) are now correctly applied
in the Bloch wave structure factor calculation using the directional DWF
formula exp(-0.5*(2pi)^2 * sum_a sa^2 * ga^2) rather than the isotropic
approximation.

- calculate_scattering_factors now takes g_vec (shape (N_g,3) or (N_g,))
  and computes the anisotropic DWF when full Cartesian vectors are supplied;
  raises a clear error if anisotropic sigmas are paired with plain magnitudes
- calculate_structure_factors computes g_vec internally from hkl and the cell,
  so the anisotropic path is always taken automatically
- StructureFactor.calculate_scattering_factors (method) updated to pass g_vec
- validate_sigmas and validate_per_atom_property moved from
  abtem.inelastic.phonons to abtem.atoms, removing the Bloch wave module's
  dependency on the frozen-phonon machinery
- A bare 3-tuple (sx, sy, sz) is now accepted as thermal_sigma, following
  the FrozenPhonons.sigmas convention (identical anisotropic sigma for all atoms)
- phonons.py re-exports the moved symbols for backward compatibility
